feat: fix and optimise db migration check
Previously, the check to skip comprehensive migration validation when there have been no changes was not working properly, since the order of CREATE INDEX statements in the DDL produced by SQLAlchemy was not deterministic.

This change sorts the output so the content is consistent if there are no changes, and adds a custom implementation for hashing the migration file content which is faster than the atlas validate command.